Fixed various warnings which were appearing under MacOSX.

+++ src/osgDB/Output.cpp
@@ -94,7 +94,7 @@ bool Output::getUniqueIDForObject(const osg::Object* obj,std::string& uniqueID)
 bool Output::createUniqueIDForObject(const osg::Object* obj,std::string& uniqueID)
 {
     char str[256];
-    sprintf(str,"%s_%i",obj->className(),_objectToUniqueIDMap.size());
+    sprintf(str,"%s_%i",obj->className(),(unsigned int)_objectToUniqueIDMap.size());
     uniqueID = str;
     return true;
 }
